requirements.txt
beautifulsoup4==4.4.1
dj-database-url==0.4.1
Django==1.9.6
django-bootstrap-form==3.2.1
django-debug-toolbar==1.4
django-modeltranslation==0.11
gunicorn==19.5.0
psycopg2==2.6.1
sqlparse==0.1.19
whitenoise==3.1
